move simple to stateManager
