+++ packages/common/src/index.ts
@@ -1,6 +1,5 @@
 export * from './common.js'
 export * from './enums.js'
 export * from './interfaces.js'
-export * from './state/index.js'
 export * from './types.js'
 export * from './utils.js'
