+++ packages/evm/package.json
@@ -57,6 +57,7 @@
   },
   "dependencies": {
     "@ethereumjs/common": "^4.3.0",
+    "@ethereumjs/statemanager": "^2.3.0",
     "@ethereumjs/tx": "^5.3.0",
     "@ethereumjs/util": "^9.0.3",
     "@types/debug": "^4.1.9",
