+++ packages/statemanager/src/stateManager.ts
@@ -1,4 +1,4 @@
-import { Chain, Common, OriginalStorageCache } from '@ethereumjs/common'
+import { Chain, Common } from '@ethereumjs/common'
 import { RLP } from '@ethereumjs/rlp'
 import { Trie } from '@ethereumjs/trie'
 import {
@@ -25,7 +25,13 @@ import {
 import debugDefault from 'debug'
 import { keccak256 } from 'ethereum-cryptography/keccak.js'
 
-import { AccountCache, CacheType, CodeCache, StorageCache } from './cache/index.js'
+import {
+  AccountCache,
+  CacheType,
+  CodeCache,
+  OriginalStorageCache,
+  StorageCache,
+} from './cache/index.js'
 
 import type {
   AccountFields,
