+++ packages/statemanager/src/statelessVerkleStateManager.ts
@@ -1,4 +1,3 @@
-import { OriginalStorageCache } from '@ethereumjs/common'
 import {
   Account,
   KECCAK256_NULL,
@@ -24,7 +23,13 @@ import debugDefault from 'debug'
 import { keccak256 } from 'ethereum-cryptography/keccak.js'
 
 import { AccessWitness, AccessedStateType, decodeValue } from './accessWitness.js'
-import { AccountCache, CacheType, CodeCache, StorageCache } from './cache/index.js'
+import {
+  AccountCache,
+  CacheType,
+  CodeCache,
+  OriginalStorageCache,
+  StorageCache,
+} from './cache/index.js'
 
 import type { AccessedStateWithAddress } from './accessWitness.js'
 import type { DefaultStateManager } from './stateManager.js'
